tiny_rlhf_scaffold/requirements.txt
transformers==4.40.0
bitsandbytes==0.39.0
sentence-transformers==2.2.2
scikit-learn==1.2.2
streamlit==1.25.0
accelerate==0.21.0
torch>=1.13.0
requests
pydantic
PyYAML
